Enhance AndroidWrappedKeyProvider and CipherSpec with additional constants and documentation; refine StorageEncryptionManager logic; clean up PredefinedKeyProvider and MockAES256KeyProvider imports.

+++ common4j/src/main/com/microsoft/identity/common/java/crypto/StorageEncryptionManager.java
@@ -478,7 +478,7 @@ public abstract class StorageEncryptionManager implements IKeyAccessor {
         // It does not fail fast on the first not equal byte to protect against
         // timing attack.
         for (int i = start; i < end; i++) {
-            result |= (byte) (expected[i - start] ^ encryptedBlob[i]);
+            result |= expected[i - start] ^ encryptedBlob[i];
         }
 
         if (result != 0) {
